LayoutTests/http/tests/permissions/permission-status-onchange-event-window-expected.txt

This test checks that the Permissions API on-change event works in window contexts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS permissionStatusState is "prompt"
PASS permissionStatusState is "granted"
PASS onChangeCalled became true
PASS successfullyParsed is true

TEST COMPLETE
